Fix bad refactor rename following merge with `main`
